Use uintptr_t for pointer alignment cast in pmurhash.c
Casting a pointer to 'long' truncates on LLP64 platforms (64-bit
Windows, where long is 32 bits), which clang flags on Windows ARM64:

  pmurhash.c:232:12: warning: cast to smaller integer type 'long'
  from 'const unsigned char *' [-Wpointer-to-int-cast]

The truncation is harmless in practice since only the low two bits
are used, but uintptr_t is the correct type for the computation and
silences the warning. stdint.h must be included before the uint32_t
macro renaming, and is guaranteed since R requires C99 or later.

+++ src/pmurhash.c
@@ -49,6 +49,7 @@ on big endian machines, or a byte-by-byte read if the endianess is unknown.
 
 
 #include "pmurhash.h"           /* renamed for consistency in digest's src/ */
+#include <stdint.h>             /* for uintptr_t; must precede the uint32_t rename below */
 
 /* I used ugly type names in the header to avoid potential conflicts with
  * application or system typedefs & defines. Since I'm not including any more
@@ -229,7 +230,7 @@ void PMurHash32_Process(uint32_t *ph1, uint32_t *pcarry, const void *key, int le
   /* This CPU does not handle unaligned word access */
 
   /* Consume enough so that the next data byte is word aligned */
-  int i = -(long)ptr & 3;
+  int i = (int)(-(uintptr_t)ptr & 3);
   if(i && i <= len) {
       DOBYTES(i, h1, c, n, ptr, len);
   }
